refactor: split middleware object tests into seperate files
