+++ QualityControl/lib/api.js
@@ -19,7 +19,7 @@ import { layoutOwnerMiddleware } from './middleware/layouts/layoutOwner.middlewa
 import { layoutIdMiddleware } from './middleware/layouts/layoutId.middleware.js';
 import { layoutServiceMiddleware } from './middleware/layouts/layoutService.middleware.js';
 import { statusComponentMiddleware } from './middleware/status/statusComponent.middleware.js';
-import { getObjectsValidationMiddleware } from './middleware/objects/ObjectGet.middleWare.js';
+// import { getObjectsValidationMiddleware } from './middleware/objects/ObjectGet.middleWare.js';
 
 /**
  * Adds paths and binds websocket to instance of HttpServer passed
@@ -47,18 +47,16 @@ export const setup = (http, ws) => {
     layoutRepository,
     jsonFileService,
     filterController,
+    objectGetByIdValidation,
+    objectsGetValidation,
+    objectGetContentsValidation,
   } = setupQcModel();
   statusService.ws = ws;
-  const {
-    getObjectsValidator,
-    getObjectByIdValidator,
-    getObjectContentValidator,
-  } = getObjectsValidationMiddleware(filterController);
 
-  http.get('/object/:id', getObjectByIdValidator, objectController.getObjectById.bind(objectController));
-  http.get('/object', getObjectContentValidator, objectController.getObjectContent.bind(objectController));
+  http.get('/object/:id', objectGetByIdValidation, objectController.getObjectById.bind(objectController));
+  http.get('/object', objectGetContentsValidation, objectController.getObjectContent.bind(objectController));
 
-  http.get('/objects', getObjectsValidator, objectController.getObjects.bind(objectController), { public: true });
+  http.get('/objects', objectsGetValidation, objectController.getObjects.bind(objectController), { public: true });
 
   http.get('/layouts', layoutController.getLayoutsHandler.bind(layoutController));
   http.get('/layout/:id', layoutController.getLayoutHandler.bind(layoutController));
